// src/app/interfaces/questionnaire.type.ts
  export interface ApiConfiguration{
    profileData?: any;
    baseURL:string;
    userAuthToken:string;
    solutionId:string;
    solutionType:'survey' | 'observation';
    fileSizeLimit:number;
    entityType:string;
    observationId:string;
    entityId:string;
    submissionNumber:string;
    evidenceCode:string;
    index:any;
  }